Uporządkowanie kodu

Closes #49

See merge request !53

+++ gl_05/Cube.cpp
@@ -4,7 +4,7 @@
 
 namespace engine
 {
-	Cube::Cube(float size, glm::vec3 color) : Cylinder(size, size * sqrt(2) / 2, 4, color)
+	Cube::Cube(float size, glm::vec3 color) : Cylinder(size, size * static_cast<float>(sqrt(2) / 2.0f), 4, color)
 	{
 
 	}
